# Conflicts:
#	src/test/resources/functional_tests.xml

+++ src/test/java/com/softserveinc/ita/rozetka/FilterProductTest.java
@@ -9,11 +9,9 @@ import org.testng.annotations.Test;
 
 import java.util.List;
 
+import static com.softserveinc.ita.rozetka.data.Category.LAPTOPS_AND_COMPUTERS;
 import static com.softserveinc.ita.rozetka.data.Category.SMARTPHONES_TV_AND_ELECTRONICS;
 import static com.softserveinc.ita.rozetka.data.ProductFilter.*;
-import static com.softserveinc.ita.rozetka.data.Category.LAPTOPS_AND_COMPUTERS;
-import static com.softserveinc.ita.rozetka.data.ProductFilter.AVAILABLE;
-import static com.softserveinc.ita.rozetka.data.ProductFilter.WITH_BONUS;
 import static com.softserveinc.ita.rozetka.data.ProductSort.PRICE_ASCENDING;
 import static com.softserveinc.ita.rozetka.data.ProductSort.PRICE_DESCENDING;
 import static com.softserveinc.ita.rozetka.data.subcategory.LaptopsAndComputersSubcategory.TABLETS;
@@ -122,18 +120,18 @@ public class FilterProductTest extends TestRunner {
                 .getHeader()
                 .search("Xbox");
 
-        int resultsAmountAfterSearch = searchResultsPage.getResultsAmount();
+        long resultsAmountAfterSearch = searchResultsPage.getResultsAmount();
 
         var filter = searchResultsPage.getFilter();
 
-        int resultsAmountAfterFilters = filter
+        long resultsAmountAfterFilters = filter
                 .filter(List.of(WHITE_COLOR, ROZETKA_SELLER, AVAILABLE))
                 .getResultsAmount();
 
         searchResultsPage.resetFilters();
         filter.filter(MICROSOFT_BRAND);
 
-        int resultsAmountAfterResetting = searchResultsPage.getResultsAmount();
+        long resultsAmountAfterResetting = searchResultsPage.getResultsAmount();
 
         var softAssert = new SoftAssertions();
